requirements.txt
streamlit
torch
torchvision
numpy
pillow
opencv-python
realesrgan
basicsr
